Update binlog path.
This fails to run because `build` exists an a bash script in the `nunit-console` root, so trying to create a file in `build\NUnitConsole.binlog` fails.  

I missed this in my testing because I made the change as an afterthought to try and avoid needing to modify the `.gitignore` file.  `build/` is ignored by default so I figured that'd be a good place to put it without even thinking to try it out. ☹️ 

I used `build-results` to parallel `test-results` which is currently used for all test output, but we can put it wherever.  I'll update the `.gitignore` with this folder.